Add files via upload
This is my first project related to machine learning. It requires a keras_model.h5 file, containing all your image data (yes, it's an image project) and also a labels.txt file, containing all the class names, starting with the index 0. It returns 'human face detected', if it detects a face, else it displays 'searching'.